Tests/LibWeb/Text/expected/HTML/autoplay-policy-default-blocks-audio.txt

muted video autoplays: true
audible video autoplays: false
